Day 35 Rain Alert Weather API
You'll need to either hard code your own api and email/app password info, or use an .env file to store it. This project is automated through python anywhere for this project. Tkinter is used in the class with whatsapp to send text message alerts, I set it up with email instead. Will not work on most vpns it seems.